feat: move logic from datetimepicker container to inner wrapper
