+++ app/component/DatetimepickerContainer.js
@@ -1,7 +1,5 @@
 import PropTypes from 'prop-types';
-import React, { useState } from 'react';
-import moment from 'moment';
-import { intlShape } from 'react-intl';
+import React from 'react';
 import { matchShape, routerShape } from 'found';
 import debounce from 'lodash/debounce';
 import loadable from '@loadable/component';
@@ -13,151 +11,65 @@ const Datetimepicker = loadable(
   { ssr: true },
 );
 
-function getInitialTimestamp(match, realtime) {
-  const now = realtime ? null : moment().valueOf();
-  return match.location.query.time
-    ? moment(Number(match.location.query.time) * 1000).valueOf()
-    : now;
-}
-
-function getInitialDepartureOrArrival(match) {
-  return match.location.query.arriveBy ? 'arrival' : 'departure';
-}
-
 function DatetimepickerContainer({ realtime, embedWhenClosed }, context) {
   const { router, match } = context;
 
-  const [timestamp, changeTimestampState] = useState(
-    getInitialTimestamp(match, realtime),
-  );
-  const [departureOrArrival, changeDepartureOrArrival] = useState(
-    getInitialDepartureOrArrival(match),
-  );
-
-  // change url query param time
-  const setParams = debounce(params => {
-    let time;
-    let arriveBy;
-    switch (params.timestamp) {
-      case undefined:
-        time = Math.round(timestamp / 1000);
-        break;
-      case null:
-        time = undefined;
-        break;
-      default:
-        time = Math.round(params.timestamp / 1000);
-        break;
-    }
-
-    switch (params.departureOrArrival) {
-      case undefined:
-        arriveBy = departureOrArrival === 'arrival' ? true : undefined;
-        break;
-      case 'arrival':
-        arriveBy = true;
-        break;
-      case 'departure':
-        arriveBy = undefined;
-        break;
-      default:
-        break;
-    }
-
+  const setParams = debounce((time, arriveBy) => {
     replaceQueryParams(router, match, {
       time,
       arriveBy,
     });
   }, 10);
 
-  const onTimestampChange = newTimestamp => {
-    changeTimestampState(newTimestamp);
-    setParams({ timestamp: newTimestamp });
-  };
-
-  const onTimeChange = debounce(newTime => {
-    if (newTime === null) {
-      onTimestampChange(moment().valueOf());
-      return;
-    }
-    onTimestampChange(newTime);
+  const onTimeChange = (time, arriveBy) => {
+    setParams(time, arriveBy ? 'true' : undefined);
     addAnalyticsEvent({
       action: 'EditJourneyTime',
       category: 'ItinerarySettings',
       name: null,
     });
-  }, 10);
+  };
 
-  const onDateChange = debounce(newDate => {
-    if (newDate === null) {
-      onTimestampChange(moment().valueOf());
-      return;
-    }
-    onTimestampChange(newDate);
+  const onDateChange = (time, arriveBy) => {
+    setParams(time, arriveBy ? 'true' : undefined);
     addAnalyticsEvent({
       action: 'EditJourneyDate',
       category: 'ItinerarySettings',
       name: null,
     });
-  }, 10);
+  };
 
   const onNowClick = () => {
-    changeDepartureOrArrival('departure');
-    const newTimestamp = realtime ? null : moment().valueOf();
-    changeTimestampState(newTimestamp);
-    setParams({ departureOrArrival: 'departure', timestamp: newTimestamp });
+    setParams(undefined, undefined);
   };
 
-  const onDepartureClick = () => {
-    const changes = {};
-    if (timestamp === null) {
-      const now = moment().valueOf();
-      changeTimestampState(now);
-      changes.timestamp = now;
-    }
-    if (departureOrArrival !== 'departure') {
-      changeDepartureOrArrival('departure');
-      addAnalyticsEvent({
-        event: 'sendMatomoEvent',
-        category: 'ItinerarySettings',
-        action: 'LeavingArrivingSelection',
-        name: 'SelectLeaving',
-      });
-      changes.departureOrArrival = 'departure';
-    }
-    if (changes.timestamp || changes.departureOrArrival) {
-      setParams(changes);
-    }
+  const onDepartureClick = time => {
+    setParams(time, undefined);
+    addAnalyticsEvent({
+      event: 'sendMatomoEvent',
+      category: 'ItinerarySettings',
+      action: 'LeavingArrivingSelection',
+      name: 'SelectLeaving',
+    });
   };
 
-  const onArrivalClick = () => {
-    const changes = {};
-    if (timestamp === null) {
-      const now = moment().valueOf();
-      changeTimestampState(now);
-      changes.timestamp = now;
-    }
-    if (departureOrArrival !== 'arrival') {
-      changeDepartureOrArrival('arrival');
-      addAnalyticsEvent({
-        event: 'sendMatomoEvent',
-        category: 'ItinerarySettings',
-        action: 'LeavingArrivingSelection',
-        name: 'SelectArriving',
-      });
-      changes.departureOrArrival = 'arrival';
-    }
-    if (changes.timestamp || changes.departureOrArrival) {
-      setParams(changes);
-    }
+  const onArrivalClick = time => {
+    setParams(time, 'true');
+    addAnalyticsEvent({
+      event: 'sendMatomoEvent',
+      category: 'ItinerarySettings',
+      action: 'LeavingArrivingSelection',
+      name: 'SelectArriving',
+    });
   };
 
   return (
     <Datetimepicker
-      timestamp={timestamp}
+      realtime={realtime}
+      initialTimestamp={match.location.query.time}
+      initialArriveBy={match.location.query.arriveBy === 'true'}
       onTimeChange={onTimeChange}
       onDateChange={onDateChange}
-      departureOrArrival={departureOrArrival}
       onNowClick={onNowClick}
       onDepartureClick={onDepartureClick}
       onArrivalClick={onArrivalClick}
@@ -176,7 +88,6 @@ DatetimepickerContainer.defaultProps = {
 };
 
 DatetimepickerContainer.contextTypes = {
-  intl: intlShape.isRequired,
   router: routerShape.isRequired,
   match: matchShape.isRequired,
   config: PropTypes.object.isRequired,
